pyproject.toml
[project]
name = "json-serde-exp"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "dhilipsiva", email = "dhilipsiva@pm.me" }
]
dependencies = [
    "django>=5.0.6",
    "graphene-django>=3.2.1",
    "faker>=25.3.0",
    "pytz>=2024.1",
    "orjson>=3.10.3",
    "strawberry-graphql>=0.233.1",
    "strawberry-graphql-django>=0.42.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "python-lsp-server>=1.11.0",
    "black>=24.4.2",
    "ipython>=8.24.0",
    "ipdb>=0.13.13",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["json_serde_exp"]
